requirements.txt
redis==3.5.3
Django==3.1.6
jieba==0.42.1
Whoosh==2.7.4
PyMySQL==1.0.2
django_haystack==3.0
django-simpleui==2021.2